# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AI Voice Audit API",
    "description": "\nAI Voice Audit API provides a robust set of endpoints for automating the quality assurance process in call centers using advanced AI models.\n\n## Key Features\n- **Asynchronous Audio Processing**: Upload call recordings for transcription and AI-powered evaluation.\n- **Dynamic LLM Selection**: Configure specific AI models per Line of Business (LOB).\n- **Multi-Tenancy**: Secure data isolation and tenant-specific configurations.\n- **Advanced Analytics**: Real-time dashboards and exportable reports.\n\n## Authentication\nThis API uses **JWT tokens** for secure authentication. \n1. Obtain tokens via the `/api/auth/login` endpoint.\n2. Include the access token in the `Authorization` header for all protected endpoints:\n   `Authorization: Bearer <your_access_token>`\n",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/reports/full-report": {
      "get": {
        "tags": [
          "reports"
        ],
        "summary": "Export Full Evaluation Report",
        "description": "Export a comprehensive report containing all evaluation records and AI analysis details.\n\n## Example curl:\n```bash\ncurl -H \"Authorization: Bearer <token>\" \"http://localhost:5000/api/reports/full-report?format=xlsx\" --output report.xlsx\n```",
        "operationId": "download_full_report_api_reports_full_report_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "date_from",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Start date (YYYY-MM-DD)",
              "examples": [
                "2024-01-01"
              ],
              "title": "Date From"
            },
            "description": "Start date (YYYY-MM-DD)"
          },
          {
            "name": "date_to",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "End date (YYYY-MM-DD)",
              "examples": [
                "2024-01-31"
              ],
              "title": "Date To"
            },
            "description": "End date (YYYY-MM-DD)"
          },
          {
            "name": "lob_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by LOB ID",
              "examples": [
                1
              ],
              "title": "Lob Id"
            },
            "description": "Filter by LOB ID"
          },
          {
            "name": "format",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Export format: csv, xlsx, pdf",
              "default": "csv",
              "title": "Format"
            },
            "description": "Export format: csv, xlsx, pdf"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "401": {
            "description": "Unauthorized - Authentication required or invalid token",
            "content": {
              "application/json": {
                "example": {
                  "detail": "Could not validate credentials"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - Insufficient permissions",
            "content": {
              "application/json": {
                "example": {
                  "detail": "Insufficient permissions to access this resource"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/reports/summary-report": {
      "get": {
        "tags": [
          "reports"
        ],
        "summary": "Export Summary Report",
        "description": "Export a high-level summary report with aggregated metrics.",
        "operationId": "download_summary_report_api_reports_summary_report_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "date_from",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Start date",
              "title": "Date From"
            },
            "description": "Start date"
          },
          {
            "name": "date_to",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "End date",
              "title": "Date To"
            },
            "description": "End date"
          },
          {
            "name": "lob_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by LOB",
              "title": "Lob Id"
            },
            "description": "Filter by LOB"
          },
          {
            "name": "format",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Export format: csv, xlsx, pdf",
              "default": "csv",
              "title": "Format"
            },
            "description": "Export format: csv, xlsx, pdf"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "401": {
            "description": "Unauthorized - Authentication required or invalid token",
            "content": {
              "application/json": {
                "example": {
                  "detail": "Could not validate credentials"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - Insufficient permissions",
            "content": {
              "application/json": {
                "example": {
                  "detail": "Insufficient permissions to access this resource"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/reports/detailed-ai-analysis": {
      "get": {
        "tags": [
          "reports"
        ],
        "summary": "Export Detailed AI Report",
        "description": "Export a CSV report specifically focused on AI reasoning and raw predictions.",
        "operationId": "download_detailed_ai_report_api_reports_detailed_ai_analysis_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "date_from",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Date From"
            }
          },
          {
            "name": "date_to",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Date To"
            }
          },
          {
            "name": "lob_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Lob Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "401": {
            "description": "Unauthorized - Authentication required or invalid token",
            "content": {
              "application/json": {
                "example": {
                  "detail": "Could not validate credentials"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/reports/ai-performance": {
      "get": {
        "tags": [
          "reports"
        ],
        "summary": "Export AI Performance Report",
        "description": "Export a report detailing cases where manual QA corrected AI predictions.",
        "operationId": "download_ai_performance_report_api_reports_ai_performance_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "date_from",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Date From"
            }
          },
          {
            "name": "date_to",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Date To"
            }
          },
          {
            "name": "lob_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Lob Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "401": {
            "description": "Unauthorized - Authentication required or invalid token",
            "content": {
              "application/json": {
                "example": {
                  "detail": "Could not validate credentials"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/reports/pii-audit": {
      "get": {
        "tags": [
          "reports"
        ],
        "summary": "Export PII Audit Report",
        "description": "Export a report of all PII redaction activities for compliance auditing.",
        "operationId": "download_pii_audit_report_api_reports_pii_audit_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "date_from",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Date From"
            }
          },
          {
            "name": "date_to",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Date To"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "401": {
            "description": "Unauthorized - Authentication required or invalid token",
            "content": {
              "application/json": {
                "example": {
                  "detail": "Could not validate credentials"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "api/auth/login"
          }
        }
      }
    }
  },
...
}